Fixing up check_ messages, and fixing polars ci testing
