+++ tools/pythonpkg/src/CMakeLists.txt
@@ -15,6 +15,7 @@ add_library(
   pyconnection.cpp
   python_import_cache.cpp
   pyrelation.cpp
+  dataframe.cpp
   pyresult.cpp
   map.cpp
   vector_conversion.cpp
