+++ tools/pythonpkg/src/include/duckdb_python/pandas_type.hpp
@@ -1,33 +1,18 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb_python/pandas_type.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
 #pragma once
 
-#include "duckdb_python/pybind_wrapper.hpp"
 #include "duckdb/common/types.hpp"
+#include "duckdb_python/pybind_wrapper.hpp"
+#include "duckdb_python/dataframe.hpp"
 
 namespace duckdb {
-
-class DataFrame : public py::object {
-public:
-	DataFrame(const py::object &o) : py::object(o, borrowed_t {}) {
-	}
-	using py::object::object;
-
-public:
-	static bool check_(const py::handle &object) {
-		return !object.is_none();
-	}
-};
-
-class PolarsDataFrame : public py::object {
-public:
-	PolarsDataFrame(const py::object &o) : py::object(o, borrowed_t {}) {
-	}
-	using py::object::object;
-
-public:
-	static bool IsDataFrame(const py::handle &object);
-	static bool IsLazyFrame(const py::handle &object);
-};
-
 // Pandas has two different sets of types
 // NumPy dtypes (e.g., bool, int8,...)
 // Pandas Specific Types (e.g., categorical, datetime_tz,...)
